move tagutil into common
